refactor: pull the handler into its own module
